Optional - Different Strategy using AuthorizeAttribute

+++ TodoListWebApp/Controllers/UserProfileController.cs
@@ -39,17 +39,13 @@ namespace TodoListWebApp.Controllers
     //
     // WithConditionalAccess:
     //
-    // In this example we will not use automatic authentication at the controller level.
-    // If the user is not authenticated we need to manually send an authentication request with the right resource parameter.
-    // [Authorize]
+    // Use this custom attribute to ensure that the user is signed in and a token can be acquired for the provided resource.
+    // 
+    [ConditionalAccessAuthorize(Resource = Startup.graphResourceId)]
     public class UserProfileController : Controller
     {
-        private string graphResourceId = ConfigurationManager.AppSettings["ida:GraphResourceId"];
-        private string graphUserUrl = ConfigurationManager.AppSettings["ida:GraphUserUrl"];
-        private const string TenantIdClaimType = "http://schemas.microsoft.com/identity/claims/tenantid";
-        private static string clientId = ConfigurationManager.AppSettings["ida:ClientId"];
-        private static string appKey = ConfigurationManager.AppSettings["ida:AppKey"];
-
+        private const string graphUserUrl = "https://graph.windows.net/{0}/me?api-version=2013-11-08";
+        
         //
         // GET: /UserProfile/
         public async Task<ActionResult> Index()
@@ -60,37 +56,13 @@ namespace TodoListWebApp.Controllers
             UserProfile profile;
             AuthenticationResult result = null;
 
-            //
-            // WithConditionalAccess:
-            //
-            // Check if the user is authenticated.  If not, issue a challenge for this resource.
-            //
-            if (ClaimsPrincipal.Current.Identity.IsAuthenticated == false)
-            {
-                HttpContext.GetOwinContext().Authentication.Challenge(
-                    new Microsoft.Owin.Security.AuthenticationProperties(
-                        new Dictionary<string, string>
-                        {
-                                {"ResourceId", graphResourceId}
-                        }),
-                    OpenIdConnectAuthenticationDefaults.AuthenticationType);
-
-                profile = new UserProfile();
-                profile.DisplayName = " ";
-                profile.GivenName = " ";
-                profile.Surname = " ";
-                ViewBag.ErrorMessage = "AuthorizationRequired";
-
-                return View(profile);
-            }
-
             try
             {
-                string tenantId = ClaimsPrincipal.Current.FindFirst(TenantIdClaimType).Value;
-                string userObjectID = ClaimsPrincipal.Current.FindFirst("http://schemas.microsoft.com/identity/claims/objectidentifier").Value;
+                string tenantId = ClaimsPrincipal.Current.FindFirst(Startup.TenantIdClaimType).Value;
+                string userObjectID = ClaimsPrincipal.Current.FindFirst(Startup.ObjectIdClaimType).Value;
                 AuthenticationContext authContext = new AuthenticationContext(Startup.Authority, new NaiveSessionCache(userObjectID));
-                ClientCredential credential = new ClientCredential(clientId, appKey);
-                result = authContext.AcquireTokenSilent(graphResourceId, credential, new UserIdentifier(userObjectID, UserIdentifierType.UniqueId));
+                ClientCredential credential = new ClientCredential(Startup.clientId, Startup.appKey);
+                result = authContext.AcquireTokenSilent(Startup.graphResourceId, credential, new UserIdentifier(userObjectID, UserIdentifierType.UniqueId));
 
                 //
                 // Call the Graph API and retrieve the user's profile.
@@ -115,60 +87,37 @@ namespace TodoListWebApp.Controllers
                 else
                 {
                     //
-                    // If the call failed, then drop the current access token and show the user an error indicating they might need to sign-in again.
+                    // If the call failed due to authorization, then drop the current access token and show the user an error indicating they might need to sign-in again.
                     //
-                    var todoTokens = authContext.TokenCache.ReadItems().Where(a => a.Resource == graphResourceId);
-                    foreach (TokenCacheItem tci in todoTokens)
-                        authContext.TokenCache.DeleteItem(tci);
+                    if (response.StatusCode == System.Net.HttpStatusCode.Unauthorized)
+                    {
+                        var todoTokens = authContext.TokenCache.ReadItems().Where(a => a.Resource == Startup.graphResourceId);
+                        foreach (TokenCacheItem tci in todoTokens)
+                            authContext.TokenCache.DeleteItem(tci);
 
-                    profile = new UserProfile();
-                    profile.DisplayName = " ";
-                    profile.GivenName = " ";
-                    profile.Surname = " ";
-                    ViewBag.ErrorMessage = "UnexpectedError";
+                        profile = new UserProfile();
+                        profile.DisplayName = " ";
+                        profile.GivenName = " ";
+                        profile.Surname = " ";
+                        ViewBag.ErrorMessage = "AuthorizationRequired";
+                        return View(profile);
+                    }
                 }
 
-                return View(profile);
             }
-            catch (AdalSilentTokenAcquisitionException ee)
+            catch (Exception ex)
             {
                 //
-                // If the user doesn't have an access token, they need to re-authorize.
+                // If the call failed for any other reason, show the user an error.
                 //
-
-                //
-                // If refresh is set to true, the user has clicked the link to be authorized again.
-                //
-                if (Request.QueryString["reauth"] == "True")
-                {
-                    //
-                    // Send an OpenID Connect sign-in request to get a new set of tokens.
-                    // If the user still has a valid session with Azure AD, they will not be prompted for their credentials.
-                    // The OpenID Connect middleware will return to this controller after the sign-in response has been handled.
-                    //
-                    // WithConditionalAcccess:  Include the resource ID for which the token is being requested.
-                    //
-                    HttpContext.GetOwinContext().Authentication.Challenge(
-                        new Microsoft.Owin.Security.AuthenticationProperties(
-                            new Dictionary<string, string>
-                            {
-                                {"ResourceId", graphResourceId }
-                            }),
-                        OpenIdConnectAuthenticationDefaults.AuthenticationType);
-                }
-
-                //
-                // The user needs to re-authorize.  Show them a message to that effect.
-                //
-                profile = new UserProfile();
-                profile.DisplayName = " ";
-                profile.GivenName = " ";
-                profile.Surname = " ";
-                ViewBag.ErrorMessage = "AuthorizationRequired";
-
-                return View(profile);
-
             }
+
+            profile = new UserProfile();
+            profile.DisplayName = " ";
+            profile.GivenName = " ";
+            profile.Surname = " ";
+            ViewBag.ErrorMessage = "UnexpectedError";
+            return View(profile);
         }
     }
 }
